feat : Venue 관련 domain 수정
# #############################
# # 타입 설정                  ##
# #############################

# # feat: 기능 구현
# # refactor: 코드 구조 개선, 변수명 변경 등 코드 리팩토링
# # fix: 버그 또는 오류 해결
# # style: CSS 등 스타일 관련 변경
# # test: 테스트 코드 추가 또는 수정
# # docs: README.md 등 문서 관련 변경
# # chore: 파일 또는 디렉토리 구조 변경, 패키지 설치 등 빌드 시스템이나 외부 도구 관련 작업
# # build: CI/CD 관련 변경
