+++ showhive-domain/domain/src/main/java/com/showhive/venue/domain/Seat.java
@@ -1,71 +1,24 @@
 package com.showhive.venue.domain;
 
-import com.showhive.BaseEntity;
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.Table;
-import jakarta.validation.constraints.NotNull;
+import com.showhive.venue.entity.SeatType;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-import java.util.Objects;
-
-@Table(name = "seats")
-@Entity
+/**
+ * 좌석 정보
+ */
 @Getter
 @Builder
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor(access = AccessLevel.PROTECTED)
-public class Seat extends BaseEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "seat_id")
-    private Long id;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "venue_id")
-    private Venue venue;
-
-    private String seatColumn;
-
-    private Short seatRow;
-
-    private Short seatFloor;
-
-    @NotNull
-    @Enumerated(EnumType.STRING)
-    private SeatType seatType;
-
-    @ManyToOne
-    @JoinColumn(name = "seat_grade_id")
-    private SeatGrade seatGrade;
-
-    public static Seat create(Venue venue, String seatColumn,
-                              Short seatRow, Short seatFloor,
-                              SeatType seatType, SeatGrade seatGrade) {
-        return Seat.builder()
-                .venue(venue)
-                .seatColumn(seatColumn)
-                .seatRow(seatRow)
-                .seatFloor(seatFloor)
-                .seatType(seatType)
-                .seatGrade(seatGrade)
-                .build();
-    }
-
-    public boolean isInVenue(Venue venue) {
-        return Objects.equals(this.getVenue().getId(), venue.getId());
-    }
+public class Seat {
+    private SeatId id;
+    private String column;
+    private Short row;
+    private Short floor;
+    private SeatGrade grade;
+    private SeatType type;
 }
