+++ showhive-domain/domain/src/main/java/com/showhive/venue/domain/SeatId.java
@@ -1,0 +1,10 @@
+package com.showhive.venue.domain;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class SeatId {
+    private Long seatId;
+}
