+++ showhive-domain/domain/src/main/java/com/showhive/venue/domain/VenueId.java
@@ -1,0 +1,4 @@
+package com.showhive.venue.domain;
+
+public record VenueId(Long id) {
+}
