+++ showhive-domain/domain/src/main/java/com/showhive/venue/domain/Zone.java
@@ -1,29 +1,28 @@
 package com.showhive.venue.domain;
 
-import com.showhive.BaseEntity;
 import java.util.ArrayList;
 import java.util.List;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
+import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+/**
+ *
+ */
 @Getter
+@EqualsAndHashCode
 @Builder
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor(access = AccessLevel.PROTECTED)
-public class Venue extends BaseEntity {
-    private VenueId id;
+public class Zone {
+    private final List<Seat> seats = new ArrayList<>();
+    private ZoneId zoneId;
     private String name;
-    private String address;
-    private Double latitude;
-    private Double longitude;
-    private String contactNumber;
-    private String link;
-    private List<Zone> zones = new ArrayList<>();
 
-    public void addZone(Zone zone) {
-        zones.add(zone);
+    public void addSeat(Seat seat) {
+        seats.add(seat);
     }
 }
