+++ showhive-domain/domain/src/main/java/com/showhive/venue/domain/ZoneId.java
@@ -1,0 +1,4 @@
+package com.showhive.venue.domain;
+
+public record ZoneId(Long id) {
+}
